fix: restructure folder
