+++ app/_components/sections/Home/AdvantagesSection/index.tsx
@@ -1,0 +1,3 @@
+export const AdvantagesSection = () => {
+  return <div>AdvantagesSection</div>;
+};
