+++ app/_components/sections/Home/index.ts
@@ -1,0 +1,2 @@
+export * from "./HeroSection";
+export * from "./DestinationCountriesSection";
